code for creating molecular graph datasets
This file contains the code for creating datasets of (molecular graph, affinity measure/label) tuples from (SMILES, affinity measure/label) tuples. RDKit was used in conjunction with the DGL-LifeSci library to create and featurize the molecular graphs. In addition, meta-learning dataset construction must be done in a certain way, with training from various tasks being interleaved so as to be converge on a shared representation. Code for doing this is also part of this file as the MetaDataClass class.